finished part 3
