+++ NetworkBasics/app/src/main/java/com/thadocizn/networkbasics/Constants.java
@@ -1,0 +1,10 @@
+package com.thadocizn.networkbasics;
+
+public class Constants {
+
+    public static final int DATABASE_VERSION = 2;
+    public static final String DATABASE_NAME = "comics";
+    public static final String TABLE_NAME = "comics";
+    public static final String COLUMN_NAME_TIMESTAMP = "comic_timestamp";
+    public static final String COLUMN_NAME_BOOL = "comic_bool";
+}
