+++ NetworkBasics/app/src/main/java/com/thadocizn/networkbasics/activities/MainActivity.java
@@ -1,4 +1,4 @@
-package com.thadocizn.networkbasics;
+package com.thadocizn.networkbasics.activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -7,6 +7,11 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.thadocizn.networkbasics.classes.NetworkAdapter;
+import com.thadocizn.networkbasics.R;
+import com.thadocizn.networkbasics.classes.XkcdComic;
+import com.thadocizn.networkbasics.data.XkcdDao;
 
 
 public class MainActivity extends AppCompatActivity {
