+++ NetworkBasics/app/src/main/java/com/thadocizn/networkbasics/classes/XkcdComic.java
@@ -1,4 +1,4 @@
-package com.thadocizn.networkbasics;
+package com.thadocizn.networkbasics.classes;
 
 import android.graphics.Bitmap;
